Add new JavaScript and HTML files for object challenge and class practice
Added `objects-challenge.js`, `objects-challenge.html` and `classes-practice.js` in the Javascript messing-around directory. This included defining several object models and classes in these files along with their methods and invoking them. The objects include a rabbit, spaceship, shopping cart and a lever in the objects-challenge. The classes define different types of monsters, a dragon and a werewolf. This commit covers the programming exercise requirements.