Add Library and Messing Around JavaScript files
Added two new JavaScript files: one for a library system and another for a 'messing-around' implementation. The library file includes a 'Book' class setup, basic library functions such as checkout and listing books, and a data structure to accept new books. The 'messing-around' file includes an instrument object and a console log statement for testing.